Add a sample Apache conf file and HACKING instructions
On XAMPP, the provided .htaccess file doesn't work out of the box,
and resists being used if the site is served from a subdirectory
(http://localhost/crosswalk-website for example).

I ended up using a separate configuration file on the directory
containing the project, and ignoring the .htaccess file, replacing
its rewrite rules with my own. This doesn't impact on the use
of .htaccess in the production version of the site, but makes
setup easier in environments where .htaccess isn't suitable.