Always use and write the SHA to the REVISION file
If the requested revision to change to did not include a SHA, the REVISION
file wouldn't include the SHA either. This would cause problems if a quick
fix was applied and the administrator forgot to include the SHA.

By always using the SHA, the code can now revert to older revisions on the
same branch (vs. always pulling tip from the branch)

Signed-off-by: James Ketrenos <james.p.ketrenos@intel.com>