Ensure cached wiki file name matches webhook html_url
Previously, cached wiki pages were not being removed in response
to the webhook as there was a mismatch between the filename hash
produced by the caching HTTP client and the hash produced from
the html_url in the data posted from github when the wiki changed.

This patch normalises the cache filename in both cases
(lowercase and with special characters decoded from the URL)
so that they match correctly.