Allow traversal to links which open a new window
The previous code was intercepting all link clicks to
external links (with a '#' in them) and passing them to Google
Analytics before setting the href of the window to the original
link href. This meant that any target attribute was being ignored.

Allow the default action to take place (i.e. navigation) instead,
so we can open links in new windows/tabs.