Make regen.php errors visible
On the live site, regen.php doesn't appear to be correctly updating
the wiki Pages/History pages for the site. However, I can't see
what's going on, because all the error messages are suppressed.

Capture any errors and present them in a non-threatening fashion
(sans stack trace) to the caller, so I can debug.