[Android] Deprecate xwalk_android_gyp
In M36, android_gyp is still available in upstream, but it is
going to be removed in recent future.
We need to deprecate xwalk_android_gyp together with upstream's
change, and we need to pass '-DOS=android' explicity to it to
generate android building environment.

This change is the prerequisite of the change in crosswalk:
https://github.com/crosswalk-project/crosswalk/pull/2079

BUG=https://crosswalk-project.org/jira/browse/XWALK-1803