[Android ext] Rewrite as simpler 'echo' example
As it's not possible to play audio files from file:// URIs in
versions of Crosswalk > 4 and < 8, replace the Android extensions
tutorial with a simpler "echo" tutorial.

This is also hopefully more useful to developers, as it contains
less code (generally) and should make it easier for them to
understand the parts of an extension and how they interact.