[faq] Update section about ignoring GPU blacklist
The xwalk-command-line file should end up in the assets/
directory of the apk file; however, placing this file in
the same directory as the web application no longer has the
desired effect, as the web assets are copied to assets/www. If
the xwalk-command-line file ends up in assets/www, it is ignored.

Change the advice so that the correct mechanism is explained
for each use case (Cordova, embedding, make_apk.py).

Fixes https://crosswalk-project.org/jira/browse/XWALK-2358