Building Crosswalk: Add proxy setup instructions.
Be explicit about what variables need to be set (at least on Linux) for
a source checkout to proceed correctly when one is behind a network
proxy.

Not only do we need to set the usual http{s}_proxy environment
variables, but a special Boto file also needs to be created for Google
Storage access to work.