Web APIs: Link to the Editor's Draft version of the Web Storage spec.
Blink, as well as Gecko, implements the Editor's Draft version of the
Web Storage spec (periodically synced from the WHATWG version). Since
Chromium M41, Blink has implemented a change to match the spec that
makes the behavior different from the W3C TR version we were referencing
and testing.

To prevent further confusion, link to the spec version that is actually
being implemented.

BUG=XWALK-3527