Initial Structure for Chinese Contents
In this solution we've duplicate each line of contents with its chinese
translations, and added a new attribute: lang="en" or "zh" to identify
if this content is for English or Chinese. We also created a cookie
"local_lang" to save the choice of the user, this cookie will get loaded
everytime that every page get rendererd, to update the body element
class attribute, and the replacement of this attribute can trigger the
replacement of CSS rule for the element with lang settings, which
controls if this element will be shown or not.

The benefit of this solution is that we can keep the website logic all
together in the same place for both English and Chinese contents, very
easy to maintain. And the cons is as Markdown grammar do not support to
add attributes into element, we have to replace the existing markdown
files with the filetype which support this feature, maybe jade is a best
choice.

BUG=XWALK-6350