Updated the way to detect Chinese content
Removed the 'language' attribute from the entry of _data.json in each
level of content directories, instead detect the language by the
filename. If it ends with '_zh', the content will be treated as Chinese,
and by default we take it as English.

Also updated the indentation of documentation-nav.ejs.