Use static Markdown for the Downloads page.
Replace the previous version, which used PHP and several synchronous
connections to download.01.org to retrieve up-to-date version numbers,
with a static one that is easier to maintain and does not run any PHP
code.

The downloads table has also been updated to only include stable version
numbers, which change infrequently and can be updated as part of the
release process: the betas and canaries, which change much more often,
are now just mentioned by name and users interested in them are directed
to download.01.org.