android build: Rename some sections to avoid confusion
The organization of the sections in this page has confused more than one
person: they assume "Fetching the source code" and "Dependency
installation" are subsections of "Crosswalk 22 and earlier", while they
are actually at the same level and must be run regardless of the
Crosswalk version the user wants to build.

Moving them to another location isn't really an option, as those
instructions should be followed in the current order. Add an "all
versions" to those 2 sections to make it more clear that they apply to
all Crosswalk versions.