building crosswalk: Move gyp instructions to the end of each page
More than one person was confused by our build instructions and thought
"fetching the source code" and "dependency installation" were
subsections of "Crosswalk 22 and earlier", while they are in fact all at
the same level.

Since gyp is deprecated, move that section to the bottom of each page
and just mention it at the beginning so that people wanting to build
older Crosswalk releases can follow the right instructions.